Reorganization of test cases between C# language versions
